pyproject.toml
[project]
name = "agrolidar"
version = "0.1.0"
description = "Safety-critical LiDAR perception for agricultural machinery"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24,<2.0",
    "torch>=2.2,<3.0",
    "PyYAML>=6.0,<7.0",
    "matplotlib>=3.8,<4.0",
    "tqdm>=4.66,<5.0",
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "pydantic>=2.0,<3.0",
    "structlog>=24.0",
    "slowapi>=0.1.9",
    "mlflow>=3.10,<4.0",
    "httpx>=0.27,<1.0",
    "onnx>=1.20,<2.0",
    "onnxruntime>=1.18,<2.0",
    "onnxsim>=0.4,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
    "mypy>=1.10",
]
viz = [
    "open3d>=0.18; platform_system != 'Linux' or platform_machine != 'aarch64'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E", "F", "I", "N", "UP", "ANN", "S", "B", "A"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=lidar_perception --cov-report=term-missing"
